Switch ZNSDomainProvider to hooks
