+++ src/App.tsx
@@ -33,7 +33,6 @@ import backgroundImage from 'assets/background.jpg';
 //- Page Imports
 import { ZNS, Staking } from 'pages';
 import PageContainer from 'containers/PageContainer';
-import { ZNSDomainsProvider } from 'lib/providers/ZNSDomainProvider';
 
 // Web3 library to query
 function getLibrary(provider: any): Web3Provider {
@@ -96,11 +95,9 @@ function wrappedApp() {
 						<Web3ReactProvider getLibrary={getLibrary}>
 							{/* Our Hooks  */}
 							<MvpVersionProvider>
-								<ZNSDomainsProvider>
-									<EnlistProvider>
-										<App />
-									</EnlistProvider>
-								</ZNSDomainsProvider>
+								<EnlistProvider>
+									<App />
+								</EnlistProvider>
 							</MvpVersionProvider>
 						</Web3ReactProvider>
 					</SubgraphProvider>
